feat: Allow multiple valid answers for tasks
Add support for multiple accepted correct answers in task definitions.

Introduced `StringSet` type to represent the value of `expected-result`
task property.

`expected-result` can be a string if there is only a single valid
answer or a list of strings if there are multiple possible correct
answers.
All values must follow the `response-result-format` precisely.

The task will pass if the response matches any of the accepted answers.
If there is no match, the results will show a diff for each possible
answer.